Relocated `User` and `UserManager` entities
